requirements.txt
attrs==23.1.0
click==8.1.3
colorama==0.4.6
flasgger==0.9.5
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mistune==2.0.5
mysql-connector-python==8.0.32
numpy==1.24.2
protobuf==3.20.3
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
six==1.16.0
tzdata==2023.3
Werkzeug==2.2.3
urllib3==1.26.15
dask==2023.4.0